requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
narwhals==1.32.0
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.11.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyvis==0.3.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.44.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
